+++ routes/api/contacts.js
@@ -2,7 +2,7 @@ const express = require('express');
 
 const router = express.Router();
 
-const ctrl = require('../../controllers/contacts');
+const ctrl = require('../../controllers');
 
 const { validateBody, isValidId } = require('../../middlewares');
